proj_2/PaymentService/requirements.txt
pika
retry
databases
sqlalchemy
asyncpg
psycopg2-binary
asyncio
fastapi
uvicorn
dependency-injector